Initial import from GEOSgcm_GridComp.
This import is on an orphan branch, because we do not know how to root
it in the main git repo.  The GEOS version was copied from the GFDL
version long before transitioning to git.  Manual processes have kept
the two versions "close" but not close enough to risk any direct merge
process.

Hopefully soon, the differences can be reconciled and GEOS can work on
a true branch of the GFDL repository.